Use 32-bit sizes for arrays. This is a performance increase as Buffer.BlockCopy is faster than my naive LongBlockCopy.

+++ ZfsSharpLib/HardDisk/FileHardDisk.cs
@@ -29,7 +29,7 @@ namespace ZfsSharp.HardDisks
             }
         }
 
-        public override void ReadBytes(byte[] array, long arrayOffset, long offset, long count)
+        public override void ReadBytes(byte[] array, int arrayOffset, long offset, int count)
         {
             CheckOffsets(offset, count);
             using (var s = mFile.CreateViewStream(offset, count, MemoryMappedFileAccess.Read))
